+++ index.css
@@ -1,7 +1,7 @@
 .header-container {
     display: flex;
     justify-content: space-between;
-    border: 2px solid black;
+    border: 1px solid black;
 }
 
 .logo {
