+++ index.html
@@ -14,6 +14,6 @@
         <!-- React render effects only div with id root -->
         <h1>Not rendered</h1> 
     </div>
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
 </body>
 </html>
